+++ server/src/photographers/photographers.service.ts
@@ -1,8 +1,9 @@
-import { HttpStatus, Injectable } from "@nestjs/common";
+import { HttpStatus, Injectable, NotFoundException } from "@nestjs/common";
 import { Photographer } from "./photographer.entity";
 import { InjectRepository } from "@nestjs/typeorm";
 import { In, Repository } from "typeorm";
 import { Client } from "src/clients/client.entity";
+import { CreatePhotographerDto } from "./dto/create-photographer.dto";
 
 @Injectable ()
 export class PhotographersService {
@@ -13,21 +14,31 @@ export class PhotographersService {
         private readonly photographerRepository: Repository<Photographer>
     ) {}
 
-    async create(photographerNew: Photographer): Promise<Photographer> {
+    async create(photographerNew: CreatePhotographerDto): Promise<Photographer> {
         const photographer = this.photographerRepository.create();
         photographer.fullname = photographerNew.fullname;
         photographer.phone = photographerNew.phone;
         photographer.work_exp = photographerNew.work_exp;
-        photographer.clients = null;
         await this.photographerRepository.save(photographer);
         return photographer;
     }
 
     async findOne(id: number): Promise<Photographer> {
-        return this.photographerRepository.findOne({
-            where: { id },
-            relations: {clients: true}
-        });
+        try{
+            const photographer = await this.photographerRepository.findOne({
+                where: { id },
+                relations: {clients: true}
+            });
+
+            if(!photographer){
+                throw new NotFoundException(`Фотографа с id ${id} не найдено`);
+            }
+
+            return photographer;
+        }
+        catch(error){
+            throw error;
+        }
     }
 
     async findAll(): Promise<Photographer[]> {
@@ -36,20 +47,48 @@ export class PhotographersService {
     }
 
     async update(id: number, updatedPhotographer: Photographer): Promise<Photographer>  {
-        const photographer = await this.photographerRepository.findOne({where: { id }});
-        photographer.fullname = updatedPhotographer.fullname;
-        photographer.phone = updatedPhotographer.phone;
-        photographer.work_exp = updatedPhotographer.work_exp;
-        const clients = await this.clientRepository.findBy({
-            id: In(updatedPhotographer.clients),
-        });
-        photographer.clients = clients;
-        await this.photographerRepository.save(photographer);
-        return photographer;
+
+        try{
+            const photographer = await this.photographerRepository.findOne({
+                where: { id },
+                relations: {clients: true}
+            });
+
+            if(!photographer){
+                throw new NotFoundException(`Фотографа с id ${id} не найдено`);
+            }
+
+            photographer.fullname = updatedPhotographer.fullname;
+            photographer.phone = updatedPhotographer.phone;
+            photographer.work_exp = updatedPhotographer.work_exp;
+            const clients = await this.clientRepository.findBy({
+                id: In(updatedPhotographer.clients),
+            });
+            photographer.clients = clients;
+            await this.photographerRepository.save(photographer);
+            return photographer;
+        }
+        catch(error){
+            throw error;
+        }
     }
 
     async remove(id: number) {
-        this.photographerRepository.delete({id});
-        return HttpStatus.OK;
+        try{
+            const photographer = await this.photographerRepository.findOne({
+                where: { id },
+                relations: {clients: true}
+            });
+
+            if(!photographer){
+                throw new NotFoundException(`Фотографа с id ${id} не найдено`);
+            }
+
+            this.photographerRepository.delete({id});
+            return HttpStatus.OK;
+        }
+        catch(error){
+            throw error;
+        }
     }
 }
